Reworked cafe update endpoint
